requirements.txt
transformers
datasets
pandas
pyyaml
jinja2
torch
black
jupyter
tokreate
evaluate
nltk
absl-py
rouge_score
bioc
huggingface-hub # For authentication of private / restricted repos
seaborn
matplotlib
jsonschema
tabulate
tiktoken
thefuzz
jsonformer
beaker-py
mergekit
pytest
openpyxl
bert_score
spacy
